Add small test framework
The intention is to be able to cut the dependency to the old version of
cxxtest that we are using.

This test framework is mainly optimised for simplicity, not comfort. It
offers a set of assertions (some were previously present in the Assert
class) and a smallish command-line test driver, and that's about it.